+++ contracts/UpdateOrganization.sol
@@ -26,8 +26,8 @@ contract UpdateOrganization is ExtraordinaryGA {
         // ⅔ have to vote “yes”
         // for * 3 >= (for + against) * 2
         Proposal storage proposal = proposals[UPDATE_ORGANIZATION][proposalId];
-        proposal.result = proposal.votesFor.mul(uint(3)) >=
-            proposal.votesFor.add(proposal.votesAgainst).mul(uint(2));
+        proposal.result = proposal.votesFor.mul(uint256(3)) >=
+            proposal.votesFor.add(proposal.votesAgainst).mul(uint256(2));
     }
 
 }
